flatten server directory structure
