+++ server/app/__init__.py
@@ -1,0 +1,14 @@
+__all__ = [
+    'admin',
+    'backends',
+    'fixtures'
+    'migrations',
+    'models',
+    'schema',
+    'signals'
+    'tests',
+    'views',
+    'apps'
+]
+
+default_app_config = 'app.apps.CruxAppConfig'
