+++ server/app/admin/__init__.py
@@ -1,0 +1,6 @@
+from .analysis import AnalysisAdmin
+from .dataset import DatasetAdmin
+from .file import FileAdmin
+from .task import TaskAdmin
+from .user import UserAdmin
+from .activity import AnalysisActivityAdmin
