+++ server/app/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class CruxAppConfig(AppConfig):
+    name = 'app'
+    verbose_name = 'CRUX'
+
+    def ready(self):
+        import app.signals
