+++ server/app/signals/__init__.py
@@ -1,0 +1,2 @@
+from .activity import (log_analysis_activity,
+                       log_dataset_activity, log_task_activity)
