--- server/crux_app/__init__.py
@@ -1,13 +1,0 @@
-__all__ = [
-    'admin',
-    'backends',
-    'fixtures'
-    'migrations',
-    'models',
-    'schema',
-    'tests',
-    'views',
-    'apps'
-]
-
-default_app_config = 'crux_app.apps.CruxAppConfig'
