--- server/crux_app/admin/__init__.py
@@ -1,5 +1,0 @@
-from .analysis import AnalysisAdmin
-from .dataset import DatasetAdmin
-from .file import FileAdmin
-from .task import TaskAdmin
-from .user import UserAdmin
