--- server/crux_app/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CruxAppConfig(AppConfig):
-    name = 'crux_app'
-    verbose_name = 'CRUX'
