--- server/crux_app/fixtures/users.yaml
@@ -1,26 +1,0 @@
-- model: crux_app.user
-  pk: 1
-  fields: 
-    email: sab.natarajan@gmail.com
-    password: pbkdf2_sha256$100000$fgKEIZHhv3rh$vWyZQAxVZMJtiO2jRQ5u/VThuVvwL2mSZ1ceJ1eLRrg=
-    first_name: Sabareesh
-    last_name: Natarajan
-    username: sabnatarajan
-    is_staff: true
-    is_superuser: true
-
-- model: crux_app.user
-  pk: 2
-  fields:
-    email: sn2744@gmail.com
-    password: pbkdf2_sha256$100000$fgKEIZHhv3rh$vWyZQAxVZMJtiO2jRQ5u/VThuVvwL2mSZ1ceJ1eLRrg=
-    first_name: Sab
-    last_name: Natarajan
-    username: sn2744
-
-- model: crux_app.user
-  pk: 3
-  fields:
-    email: t3kcit@gmail.com
-    password: pbkdf2_sha256$100000$A8c6mBGM2W6c$vVv2kHriMdFiDyMcdtbLBM0ifFhuFyOapS3ELku4p8I=
-    username: t3kcit
