+++ server/manage.py
@@ -4,7 +4,7 @@ import sys
 
 if __name__ == "__main__":
     DJANGO_SETTINGS_MODULE = os.environ.get(
-        'DJANGO_SETTINGS_MODULE') or 'crux.settings.development'
+        'DJANGO_SETTINGS_MODULE') or 'settings.development'
     os.environ.setdefault("DJANGO_SETTINGS_MODULE",
                           DJANGO_SETTINGS_MODULE)
     try:
